Added implementation files for inline methods Edited code to fix linting Addressed review comments
