+++ rmw_connext_cpp/include/rmw_connext_cpp/connext_static_subscriber_info.hpp
@@ -41,7 +41,17 @@ struct ConnextStaticSubscriberInfo : ConnextCustomEventInfo
   DDS::ReadCondition * read_condition_;
   bool ignore_local_publications;
   const message_type_support_callbacks_t * callbacks_;
-  rmw_ret_t get_status(const DDS_StatusMask mask, void * event) override;
+  /// Remap the specific RTI Connext DDS DataReader Status to a generic RMW status type.
+  /**
+   * @param mask input status mask
+   * @param event
+   */
+  rmw_ret_t get_status(DDS_StatusMask mask, void * event) override;
+  /// Return the topic reader entity for this subsciber
+  /**
+   *
+   * @return the topic reader associated with this subscriber
+   */
   DDSEntity * get_entity() override;
 };
 }  // extern "C"
@@ -64,67 +74,5 @@ public:
 private:
   std::atomic<std::size_t> current_count_;
 };
-
-/**
- * Remap the specific RTI Connext DDS DataReader Status to a generic RMW status type.
- *
- * @param mask input status mask
- * @param event
- */
-inline rmw_ret_t ConnextStaticSubscriberInfo::get_status(
-  const DDS_StatusMask mask,
-  void * event)
-{
-  switch (mask) {
-    case DDS_StatusKind::DDS_LIVELINESS_CHANGED_STATUS: {
-        DDS_LivelinessChangedStatus liveliness_changed;
-        DDS_ReturnCode_t dds_return_code = topic_reader_
-          ->get_liveliness_changed_status(liveliness_changed);
-
-        rmw_ret_t from_dds = check_dds_ret_code(dds_return_code);
-        if (from_dds != RMW_RET_OK) {
-          return from_dds;
-        }
-
-        rmw_liveliness_changed_status_t * rmw_liveliness_changed_status =
-          static_cast<rmw_liveliness_changed_status_t *>(event);
-        rmw_liveliness_changed_status->alive_count = liveliness_changed.alive_count;
-        rmw_liveliness_changed_status->not_alive_count = liveliness_changed.not_alive_count;
-        rmw_liveliness_changed_status->alive_count_change = liveliness_changed.alive_count_change;
-        rmw_liveliness_changed_status->not_alive_count_change =
-          liveliness_changed.not_alive_count_change;
-
-        break;
-      }
-    case DDS_StatusKind::DDS_REQUESTED_DEADLINE_MISSED_STATUS: {
-        DDS_RequestedDeadlineMissedStatus requested_deadline_missed;
-        DDS_ReturnCode_t dds_return_code = topic_reader_
-          ->get_requested_deadline_missed_status(requested_deadline_missed);
-
-        rmw_ret_t from_dds = check_dds_ret_code(dds_return_code);
-        if (from_dds != RMW_RET_OK) {
-          return from_dds;
-        }
-
-        rmw_requested_deadline_missed_status_t * rmw_requested_deadline_missed_status =
-          static_cast<rmw_requested_deadline_missed_status_t *>(event);
-        rmw_requested_deadline_missed_status->total_count = requested_deadline_missed.total_count;
-        rmw_requested_deadline_missed_status->total_count_change =
-          requested_deadline_missed.total_count_change;
-
-        break;
-      }
-
-    default:
-      return RMW_RET_UNSUPPORTED;
-  }
-  return RMW_RET_OK;
-}
-
-inline DDSEntity * ConnextStaticSubscriberInfo::get_entity()
-{
-  return topic_reader_;
-}
-
 
 #endif  // RMW_CONNEXT_CPP__CONNEXT_STATIC_SUBSCRIBER_INFO_HPP_
