+++ rmw_connext_shared_cpp/include/rmw_connext_shared_cpp/connext_static_event_info.hpp
@@ -34,25 +34,6 @@ typedef struct ConnextCustomEventInfo
   virtual rmw_ret_t get_status(const DDS_StatusMask mask, void * event) = 0;
   virtual DDSEntity * get_entity() = 0;
 
-  /**
-    * Assign the input DDS return code to its corresponding RMW return code.
-    * @param dds_return_code input DDS return code
-    * @return to_return the corresponding rmw_ret_t that maps to the input DDS_ReturnCode_t. By
-    * default RMW_RET_ERROR is returned if no corresponding rmw_ret_t is not defined.
-    */
-  rmw_ret_t check_dds_ret_code(DDS_ReturnCode_t & dds_return_code)
-  {
-    switch (dds_return_code) {
-      case DDS_RETCODE_OK:
-        return RMW_RET_OK;
-      case DDS_RETCODE_ERROR:
-        return RMW_RET_ERROR;
-      case DDS_RETCODE_TIMEOUT:
-        return RMW_RET_TIMEOUT;
-      default:
-        return RMW_RET_ERROR;
-    }
-  }
 } ConnextCustomEventInfo;
 
 #endif  // RMW_CONNEXT_SHARED_CPP__CONNEXT_STATIC_EVENT_INFO_HPP_
