+++ rmw_connext_shared_cpp/include/rmw_connext_shared_cpp/event_converter.hpp
@@ -15,11 +15,36 @@
 #ifndef RMW_CONNEXT_SHARED_CPP__EVENT_CONVERTER_HPP_
 #define RMW_CONNEXT_SHARED_CPP__EVENT_CONVERTER_HPP_
 
-#include "ndds/ndds_cpp.h"
-#include "rmw/event.h"
 
+#include "rmw/event.h"
+#include "rmw/ret_types.h"
+#include "rmw/types.h"
+
+#include "ndds/ndds_cpp.h"
+#include "ndds/ndds_namespace_cpp.h"
+
+#include "ndds_include.hpp"
+
+/// Return the corresponding DDS_StatusKind to the input RMW_EVENT
+/**
+ * @param event_t
+ * @return
+ */
 DDS_StatusMask get_status_mask_from_rmw(const rmw_event_type_t & event_t);
 
+/// Return true if the input RMW event has a corresponding DDS_StatusKind.
+/**
+ * @param event_t input rmw event to check
+ * @return true if there is an RMW to DDS_StatusKind mapping, false otherwise
+ */
 bool is_event_supported(const rmw_event_type_t & event_t);
+
+/// Assign the input DDS return code to its corresponding RMW return code.
+/**
+  * @param dds_return_code input DDS return code
+  * @return to_return the corresponding rmw_ret_t that maps to the input DDS_ReturnCode_t. By
+  * default RMW_RET_ERROR is returned if no corresponding rmw_ret_t is not defined.
+  */
+rmw_ret_t check_dds_ret_code(DDS_ReturnCode_t & dds_return_code);
 
 #endif  // RMW_CONNEXT_SHARED_CPP__EVENT_CONVERTER_HPP_
